requirements.txt

flask==3.0.0
flask-pymongo==2.3.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
pyotp==2.9.0
flask-limiter==3.5.0
flask-cors==4.0.0
pymongo==4.6.1
werkzeug==3.0.1